+++ Source/General/Project.vb
@@ -30,6 +30,7 @@ Public Class Project
     Public CompCheckPercentage As Double = 5.0
     Public CompCheckTestblockSeconds As Double = 2.0
     Public Compressibility As Double
+    Public ConvertChromaSubsampling As Boolean = True
     Public ConvertSup2Sub As Boolean
     Public CustomSourceDAR As String = ""
     Public CustomSourcePAR As String = ""